Apply extract class pattern to log analysis elapsed time
