+++ lib/service/database.js
@@ -1,6 +1,5 @@
 'use strict';
 
-var async = require('async');
 var dot = require('dot');
 dot.templateSettings.strip = false;
 
@@ -12,6 +11,7 @@ var debug = require('../util/debug')('database-service');
 var QueryRunner = require('../postgresql/query-runner');
 var QueryParser = require('../postgresql/query-parser');
 var BatchClient = require('../postgresql/batch-client');
+var AnalysisTimeLogger = require('./analysis-time-logger');
 
 var QUERY_RUNNER_READONLY_OP = true;
 var QUERY_RUNNER_WRITE_OP = !QUERY_RUNNER_READONLY_OP;
@@ -39,6 +39,7 @@ function DatabaseService(user, dbParams, batchParams, logger) {
     );
 
     this.logger = logger;
+    this.analysisTimeLogger = new AnalysisTimeLogger(this.logger, this.batchClient);
 }
 
 module.exports = DatabaseService;
@@ -148,90 +149,6 @@ DatabaseService.prototype.enqueue = function(queries, callback) {
         callback();
     });
 };
-
-DatabaseService.prototype.logAnalysisElapsedTime = function (rows) {
-    var self = this;
-    var logElapsedTime = self.logNodeElapsedTime.bind(self);
-    var nodesToLog = getNodesToLogFromRows(rows);
-
-    async.map(nodesToLog, logElapsedTime, function (err) {
-        if (err) {
-            return self.logger.error(err);
-        }
-    });
-};
-
-function getNodesToLogFromRows(rows) {
-    return rows
-        .map(function (row) {
-            if (row.last_modified_by) {
-                return { nodeId: row.node_id, jobId: row.last_modified_by };
-            }
-        })
-        .filter(function (toLog) {
-            return !!toLog;
-        });
-}
-
-DatabaseService.prototype.logNodeElapsedTime = function (nodeToLog, callback) {
-    var self = this;
-
-    self.getNodeElapsedTime(nodeToLog.jobId, nodeToLog.nodeId, function (err, elapsedTimeMilliseconds) {
-        if (err) {
-            return callback(err);
-        }
-
-        if (Number.isFinite(elapsedTimeMilliseconds)) {
-            self.logger.info({ node: nodeToLog.nodeId, elapsedTime: elapsedTimeMilliseconds});
-        }
-
-        callback();
-    });
-};
-
-DatabaseService.prototype.getNodeElapsedTime = function(jobId, nodeId, callback) {
-    this.batchClient.getJob(jobId, function (err, job) {
-        if (err) {
-            return callback(err);
-        }
-
-        if (!hasStartAndEndTime(job)) {
-            return callback();
-        }
-
-        var queryElapsedTime = getQueryElapsedTime(job.query.query, nodeId);
-
-        if (!queryElapsedTime) {
-            return callback();
-        }
-
-        callback(null, queryElapsedTime);
-    });
-};
-
-function hasStartAndEndTime(job) {
-    return job.status === 'done' && job.query && Array.isArray(job.query.query);
-}
-
-function getQueryElapsedTime (queries, nodeId) {
-    for (var i = 0; i < queries.length; i++) {
-        if (isQueryOfNodeId(queries[i].query, nodeId)) {
-            var start = new Date(queries[i].started_at);
-            var end = new Date(queries[i].ended_at);
-
-            if (start && end) {
-                var elapsedTimeMilliseconds = end.getTime() - start.getTime();
-
-                debug('analysis time %j', { analysis: nodeId, time: elapsedTimeMilliseconds });
-                return elapsedTimeMilliseconds;
-            }
-        }
-    }
-}
-
-function isQueryOfNodeId(query, nodeId) {
-    return query.indexOf('/* camshaft_node_id: ' + nodeId) !== -1;
-}
 
 var registerNodeQueryTemplate = dot.template([
     'SELECT',
@@ -402,7 +319,7 @@ DatabaseService.prototype.queueAnalysisOperations = function(analysis, callback)
             }
         });
 
-        self.logAnalysisElapsedTime(rows);
+        self.analysisTimeLogger.log(rows);
 
         var nodesToQueueForUpdate = sortedNodes.filter(function(node) {
             return node.isOutdated();
